# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * LifeXP API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorMessage {
  message: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface SignupRequest {
  email: string;
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
  /** @minLength 8 */
  password: string;
}

export interface UserSummary {
  id: string;
  email: string;
  username: string;
  createdAt?: string;
}

export interface SignupResponse {
  user: UserSummary;
  message: string;
}

export interface SigninRequest {
  email: string;
  password: string;
}

export interface SigninResponse {
  accessToken: string;
  refreshToken: string;
  user: UserSummary;
}

export interface TokenPair {
  accessToken: string;
  refreshToken: string;
}

export interface JwtPayload {
  sub: string;
  email: string;
  iat?: number;
  exp?: number;
}

export interface User {
  id: string;
  email: string;
  username: string;
  displayName?: string | null;
  createdAt?: string;
  lastLoginAt?: string | null;
}

export interface UpdateUserRequest {
  displayName?: string | null;
  /**
     * @minLength 3
     * @maxLength 30
     */
  username?: string;
}

export interface OnboardingStepRecord {
  id: string;
  userId: string;
  currentStep: number;
  isCompleted: boolean;
  completedAt?: string | null;
  updatedAt?: string;
}

export interface UserProfile {
  id: string;
  userId: string;
  heightCm?: number | null;
  weightKg?: string | null;
  activityLevel?: string | null;
  dateOfBirth?: string | null;
}

export interface Archetype {
  id: string;
  name: string;
  description: string;
  focusAreas?: string[];
}

export interface UserCharacter {
  id: string;
  userId: string;
  archetypeId: string;
  displayName?: string | null;
  updatedAt?: string;
}

export interface UserGoal {
  id: string;
  userId: string;
  goalKey: string;
  isPrimary: boolean;
}

export type OnboardingStateCharacter = { [key: string]: unknown } | null;

export interface OnboardingState {
  state: OnboardingStepRecord;
  profile?: UserProfile | null;
  character?: OnboardingStateCharacter;
  goals: UserGoal[];
}

export type ProfileRequestActivityLevel = typeof ProfileRequestActivityLevel[keyof typeof ProfileRequestActivityLevel] | null;


export const ProfileRequestActivityLevel = {
  sedentary: 'sedentary',
  light: 'light',
  moderate: 'moderate',
  active: 'active',
  very_active: 'very_active',
} as const;

export interface ProfileRequest {
  heightCm?: number | null;
  weightKg?: number | null;
  activityLevel?: ProfileRequestActivityLevel;
  dateOfBirth?: string | null;
}

export interface GoalsRequest {
  /** @minItems 1 */
  goals: string[];
  primaryGoal?: string;
}

export interface UserLevel {
  userId?: string;
  currentLevel: number;
  totalXp: number;
}

export type UserAttributeAttribute = typeof UserAttributeAttribute[keyof typeof UserAttributeAttribute];


export const UserAttributeAttribute = {
  STRENGTH: 'STRENGTH',
  ENDURANCE: 'ENDURANCE',
  MOBILITY: 'MOBILITY',
  NUTRITION: 'NUTRITION',
  RECOVERY: 'RECOVERY',
  DISCIPLINE: 'DISCIPLINE',
  KNOWLEDGE: 'KNOWLEDGE',
} as const;

export interface UserAttribute {
  id: string;
  userId: string;
  attribute: UserAttributeAttribute;
  currentValue: number;
}

export interface XpTransaction {
  id: string;
  userId: string;
  amount: number;
  sourceType: string;
  sourceId?: string | null;
  category?: string | null;
  description?: string | null;
  createdAt?: string;
}

export type AttributeHistoryEntryAttribute = typeof AttributeHistoryEntryAttribute[keyof typeof AttributeHistoryEntryAttribute];


export const AttributeHistoryEntryAttribute = {
  STRENGTH: 'STRENGTH',
  ENDURANCE: 'ENDURANCE',
  MOBILITY: 'MOBILITY',
  NUTRITION: 'NUTRITION',
  RECOVERY: 'RECOVERY',
  DISCIPLINE: 'DISCIPLINE',
  KNOWLEDGE: 'KNOWLEDGE',
} as const;

export interface AttributeHistoryEntry {
  id: string;
  userId: string;
  attribute: AttributeHistoryEntryAttribute;
  delta: number;
  sourceType: string;
  sourceId?: string | null;
  createdAt: string;
}

export interface ProgressionSummary {
  level: UserLevel;
  attributes: UserAttribute[];
  recentTransactions: XpTransaction[];
}

export type QuestTemplateQuestType = typeof QuestTemplateQuestType[keyof typeof QuestTemplateQuestType];


export const QuestTemplateQuestType = {
  DAILY: 'DAILY',
  WEEKLY: 'WEEKLY',
  MILESTONE: 'MILESTONE',
  CHALLENGE: 'CHALLENGE',
} as const;

export type QuestTemplateDifficulty = typeof QuestTemplateDifficulty[keyof typeof QuestTemplateDifficulty] | null;


export const QuestTemplateDifficulty = {
  EASY: 'EASY',
  MEDIUM: 'MEDIUM',
  HARD: 'HARD',
  EXPERT: 'EXPERT',
} as const;

export type QuestTemplateStatus = typeof QuestTemplateStatus[keyof typeof QuestTemplateStatus];


export const QuestTemplateStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;

export type QuestTemplateProgressionConfig = { [key: string]: unknown } | null;

export interface QuestTemplate {
  id: string;
  title: string;
  description: string;
  instructions?: string | null;
  category: string;
  questType: QuestTemplateQuestType;
  targetValue?: string | null;
  targetUnit?: string | null;
  difficulty?: QuestTemplateDifficulty;
  status: QuestTemplateStatus;
  progressionConfig?: QuestTemplateProgressionConfig;
  verificationRequirement?: string | null;
  primaryAttributes?: string[] | null;
  xpReward?: number | null;
  createdAt?: string;
}

export type UserQuestStatus = typeof UserQuestStatus[keyof typeof UserQuestStatus];


export const UserQuestStatus = {
  ASSIGNED: 'ASSIGNED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  ABANDONED: 'ABANDONED',
} as const;

export interface UserQuest {
  id: string;
  userId: string;
  questTemplateId: string;
  status: UserQuestStatus;
  progressValue: string;
  targetValue: string;
  assignedAt?: string;
  completedAt?: string | null;
}

export interface UserQuestWithTemplate {
  user_quests: UserQuest;
  quest_templates?: QuestTemplate | null;
}

export interface XpAwardResult {
  alreadyAwarded: boolean;
  xpAwarded: number;
  newLevel: number;
  leveledUp: boolean;
}

export interface QuestCompleteResponse {
  success: boolean;
  quest: UserQuest;
  xp: XpAwardResult;
  message: string;
}

/**
 * Bad request
 */
export type BadRequestResponse = ErrorMessage;

/**
 * Unauthorized
 */
export type UnauthorizedResponse = ErrorMessage;

/**
 * Not found
 */
export type NotFoundResponse = ErrorMessage;

/**
 * Conflict
 */
export type ConflictResponse = ErrorMessage;

export type RefreshTokensBody = {
  refreshToken: string;
};

export type LogoutBody = {
  refreshToken: string;
};

export type UpdateOnboardingStepBody = {
  /**
     * @minimum 1
     * @maximum 7
     */
  currentStep?: number;
  isCompleted?: boolean;
};

export type SelectArchetypeBody = {
  archetypeId: string;
};

export type GetAttributeHistoryParams = {
limit?: number;
};

export type GetRecommendedQuestsParams = {
limit?: number;
};

export type UpdateQuestProgressBody = {
  /** @minimum 0 */
  progress: number;
};

export type AbandonQuest200 = {
  success: boolean;
  quest: UserQuest;
};