# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LifeXP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quests:
    get:
      operationId: getMyQuests
      tags: [quests]
      summary: Get user's assigned quests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User quests with template info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserQuestWithTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /quests/catalogue:
    get:
      operationId: getQuestCatalogue
      tags: [quests]
      summary: Get all active quest templates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Quest templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /quests/recommended:
    get:
      operationId: getRecommendedQuests
      tags: [quests]
      summary: Get recommended quest templates not yet assigned
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Recommended templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /quests/assign/{templateId}:
    post:
      operationId: assignQuest
      tags: [quests]
      summary: Assign a quest template to the current user
      security:
        - bearerAuth: []
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Quest assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserQuestWithTemplate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /quests/{id}:
    get:
      operationId: getQuest
      tags: [quests]
      summary: Get a specific user quest
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quest with template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserQuestWithTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /quests/{id}/progress:
    patch:
      operationId: updateQuestProgress
      tags: [quests]
      summary: Update quest progress value
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                progress:
                  type: number
                  minimum: 0
              required: [progress]
      responses:
        "200":
          description: Updated quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserQuest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /quests/{id}/abandon:
    post:
      operationId: abandonQuest
      tags: [quests]
      summary: Abandon an active quest
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quest abandoned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  quest:
                    $ref: "#/components/schemas/UserQuest"
                required: [success, quest]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /quests/{id}/complete:
    post:
      operationId: completeQuest
      tags: [quests]
      summary: Complete a quest and award XP
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quest completed with XP award
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestCompleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
  schemas:
...
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    QuestTemplate:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        instructions:
          type: string
          nullable: true
        category:
          type: string
        questType:
          type: string
          enum: [DAILY, WEEKLY, MILESTONE, CHALLENGE]
        targetValue:
          type: string
          nullable: true
        targetUnit:
          type: string
          nullable: true
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD, EXPERT]
          nullable: true
        status:
          type: string
          enum: [ACTIVE, INACTIVE, ARCHIVED]
        progressionConfig:
          type: object
          nullable: true
        verificationRequirement:
          type: string
          nullable: true
        primaryAttributes:
          type: array
          items:
            type: string
          nullable: true
        xpReward:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, category, questType, status]
    UserQuest:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        questTemplateId:
          type: string
        status:
          type: string
          enum: [ASSIGNED, IN_PROGRESS, COMPLETED, ABANDONED]
        progressValue:
          type: string
        targetValue:
          type: string
        assignedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, questTemplateId, status, progressValue, targetValue]
    UserQuestWithTemplate:
      type: object
      properties:
        user_quests:
          $ref: "#/components/schemas/UserQuest"
        quest_templates:
          $ref: "#/components/schemas/QuestTemplate"
          nullable: true
      required: [user_quests]
    XpAwardResult:
      type: object
      properties:
        alreadyAwarded:
          type: boolean
        xpAwarded:
          type: integer
        newLevel:
          type: integer
        leveledUp:
          type: boolean
      required: [alreadyAwarded, xpAwarded, newLevel, leveledUp]
    QuestCompleteResponse:
      type: object
      properties:
        success:
          type: boolean
        quest:
          $ref: "#/components/schemas/UserQuest"
        xp:
          $ref: "#/components/schemas/XpAwardResult"
        message:
          type: string
      required: [success, quest, xp, message]